chat-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.352
langchain-openai==0.0.2
openai==1.3.8
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1